move HAS_UCC to _ucc_utils.py
