+++ apex/transformer/testing/commons.py
@@ -35,7 +35,7 @@ from apex.transformer.pipeline_parallel.schedules.common import (
     Batch,
 )
 from apex.transformer.testing import global_vars
-from apex.transformer.testing.distributed_test_base import HAS_UCC
+from apex.transformer._ucc_util import HAS_UCC
 
 TEST_SUCCESS_MESSAGE = ">> passed the test :-)"
