+++ tests/L0/run_transformer/run_bert_minimal_test.py
@@ -1,7 +1,7 @@
 import random
 import torch
 
-from apex.transformer.testing.distributed_test_base import HAS_UCC
+from apex.transformer._ucc_util import HAS_UCC
 from apex.transformer.enums import ModelType
 from apex.transformer import tensor_parallel
 from apex.transformer import parallel_state
